+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Next
 ----
 
 - Add support for `sphinxcontrib-mermaid`_ diagrams.
+- Add support for the ``notion-file`` directive for Notion File blocks.
 
 2026.02.09
 ----------